Port buildContext to TS
Unclutter dropdown-context
